+++ modern/org/commcare/modern/reference/JavaFileRoot.java
@@ -1,7 +1,7 @@
 /**
  *
  */
-package org.commcare.resources;
+package org.commcare.modern.reference;
 
 import org.javarosa.core.reference.PrefixedRootFactory;
 import org.javarosa.core.reference.Reference;
