+++ modern/org/commcare/modern/reference/JavaHttpReference.java
@@ -1,7 +1,7 @@
 /**
  *
  */
-package org.commcare.resources;
+package org.commcare.modern.reference;
 
 import org.javarosa.core.reference.Reference;
